Playing around with figures.
Change the position of some figures to make more sense in context of the
text. Mainly the stack frame was preceeded by two figures but both were
linked to later in the text than the stack frame one.

The preceeding indentation was removed from all code figures. Some
figures had it, some not. We removed it from all to be consistent.